move root confest to server dir
